+++ src/main/java/base/ShapeInterface.java
@@ -1,5 +1,5 @@
 package base;
 
 public interface ShapeInterface {
-    int calculateArea();
+    double calculateArea();
 }
